+++ IAT.Core/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace IAT.Core
-{
-    public class Class1
-    {
-
-    }
-}